Move ProductCard to component
