+++ portal/src/routes/(authenticated)/projects/[id=idNumber]/ProductCard.svelte
@@ -1,0 +1,226 @@
+<script lang="ts">
+  import { page } from '$app/state';
+  import Dropdown from '$lib/components/Dropdown.svelte';
+  import IconContainer from '$lib/components/IconContainer.svelte';
+  import Tooltip from '$lib/components/Tooltip.svelte';
+  import { getIcon } from '$lib/icons/productDefinitionIcon';
+  import { m } from '$lib/paraglide/messages';
+  import { getLocale, localizeHref } from '$lib/paraglide/runtime';
+  import type { ProductActionType } from '$lib/products';
+  import ProductDetails, {
+    showProductDetails
+  } from '$lib/products/components/ProductDetails.svelte';
+  import { isAdminForOrg, isSuperAdmin } from '$lib/utils/roles';
+  import { getRelativeTime } from '$lib/utils/time';
+  import { ProductType } from 'sil.appbuilder.portal.common/workflow';
+  import DeleteProduct from './modals/DeleteProduct.svelte';
+  import Properties from './modals/Properties.svelte';
+
+  type Transition = {
+    TransitionType: number;
+    InitialState: string | null;
+    WorkflowType: number | null;
+    AllowedUserNames: string | null;
+    Command: string | null;
+    Comment: string | null;
+    DateTransition: Date | null;
+    User: {
+      Name: string | null;
+    } | null;
+  };
+
+  interface Props {
+    project: {
+      Name: string | null;
+      DateArchived: Date | null;
+      Organization: {
+        Id: number;
+      };
+    };
+    product: {
+      Id: string;
+      DatePublished: Date | null;
+      DateUpdated: Date | null;
+      Properties: string | null;
+      PublishLink: string | null;
+      ProductDefinition: {
+        Name: string | null;
+        Workflow: {
+          ProductType: ProductType;
+        };
+      };
+      Store: {
+        Description: string | null;
+      } | null;
+      Transitions: Transition[];
+      UserTasks: {
+        UserId: number;
+        DateCreated: Date | null;
+      }[];
+      WorkflowInstance: unknown;
+      actions: ProductActionType[];
+      ActiveTransition?: Transition;
+      PreviousTransition?: Transition;
+    };
+  }
+
+  let { product, project }: Props = $props();
+
+  let deleteProductModal: HTMLDialogElement | undefined = $state(undefined);
+  let updateProductModal: HTMLDialogElement | undefined = $state(undefined);
+  const showTaskWaiting = $derived(!!product.WorkflowInstance);
+
+  async function handleProductAction(productId: string, action: string) {
+    try {
+      const formData = new FormData();
+      formData.append('productId', productId);
+      formData.append('productAction', action);
+
+      const response = await fetch(`${page.url.pathname}?/productAction`, {
+        method: 'POST',
+        body: formData
+      });
+
+      if (!response.ok) {
+        throw new Error('Network response was not ok');
+      }
+    } catch (error) {
+      console.error('Error performing product action:', error);
+    }
+  }
+</script>
+
+<div class="rounded-md border border-slate-400 w-full my-2">
+  <div class="bg-neutral p-2 flex flex-row rounded-t-md" class:rounded-b-md={!showTaskWaiting}>
+    <span class="grow min-w-0">
+      <IconContainer icon={getIcon(product.ProductDefinition.Name ?? '')} width="32" />
+      {product.ProductDefinition.Name}
+    </span>
+    {#if product.PublishLink}
+      {@const pType = product.ProductDefinition.Workflow.ProductType}
+      <span class="flex flex-col px-2">
+        <a class="link" href={product.PublishLink} target="_blank">
+          <IconContainer icon="ic:twotone-store" width={24} />
+        </a>
+        {#if pType !== ProductType.Web}
+          <a
+            class="link"
+            href="/api/products/{product.Id}/files/published/{pType === ProductType.AssetPackage
+              ? 'asset-package'
+              : 'apk'}"
+            target="_blank"
+          >
+            <IconContainer icon="mdi:launch" width={24} />
+          </a>
+        {/if}
+      </span>
+    {/if}
+    <span class="w-32 inline-block">
+      {m.project_products_updated()}
+      <br />
+      <Tooltip tip={product.DateUpdated?.toLocaleString(getLocale())}>
+        {getRelativeTime(product.DateUpdated)}
+      </Tooltip>
+    </span>
+    <span class="w-32 inline-block">
+      {m.project_products_published()}
+      <br />
+      <Tooltip tip={product.DatePublished?.toLocaleString(getLocale())}>
+        {getRelativeTime(product.DatePublished)}
+      </Tooltip>
+    </span>
+    <Dropdown labelClasses="px-1" contentClasses="drop-arrow bottom-12 right-0 p-1 min-w-36 w-auto">
+      {#snippet label()}
+        <IconContainer icon="charm:menu-kebab" width="20" />
+      {/snippet}
+      {#snippet content()}
+        <ul class="menu menu-compact overflow-hidden rounded-md">
+          {#each product.actions as action}
+            {@const message =
+              //@ts-expect-error this is in fact correct
+              m['products_actions_' + action]()}
+            <li class="w-full rounded-none">
+              <button
+                class="text-nowrap"
+                onclick={(event) => {
+                  handleProductAction(product.Id, action);
+                  event.currentTarget.blur();
+                }}
+              >
+                {message}
+              </button>
+            </li>
+          {/each}
+          <li class="w-full rounded-none">
+            <button class="text-nowrap" onclick={() => showProductDetails(product.Id)}>
+              {m.project_products_popup_details()}
+            </button>
+          </li>
+          <li class="w-full rounded-none">
+            <a href={localizeHref(`/products/${product.Id}/files`)} class="text-nowrap">
+              {m.project_productFiles()}
+            </a>
+          </li>
+          {#if isAdminForOrg(project.Organization.Id, page.data.session?.user.roles)}
+            <li class="w-full rounded-none">
+              <button class="text-nowrap" onclick={() => updateProductModal?.showModal()}>
+                {m.project_products_popup_properties()}
+              </button>
+            </li>
+          {/if}
+          {#if isSuperAdmin(page.data.session?.user.roles) && !!product.WorkflowInstance}
+            <li class="w-full rounded-none">
+              <a href={localizeHref(`/workflow-instances/${product.Id}`)}>
+                {m.common_workflow()}
+              </a>
+            </li>
+          {/if}
+          <li class="w-full rounded-none">
+            <button class="text-nowrap text-error" onclick={() => deleteProductModal?.showModal()}>
+              {m.models_delete({ name: m.tasks_product() })}
+            </button>
+          </li>
+        </ul>
+      {/snippet}
+    </Dropdown>
+    <DeleteProduct
+      bind:modal={deleteProductModal}
+      {product}
+      endpoint="deleteProduct"
+      project={project.Name ?? m.tasks_project()}
+    />
+    <Properties bind:modal={updateProductModal} {product} endpoint="updateProduct" />
+  </div>
+  {#if showTaskWaiting}
+    <div class="p-2 flex gap-1">
+      {#if project.DateArchived}
+        {@html m.tasks_archivedAt({
+          activityName: product.ActiveTransition?.InitialState ?? ''
+        })}
+      {:else}
+        <span class="text-red-500">
+          {m.tasks_waiting({
+            // waiting since EITHER (the last task exists) -> that task's creation time
+            // OR (there are no tasks for this product) -> the last completed transition's completion time
+            waitTime: getRelativeTime(
+              product.UserTasks.slice(-1)[0]?.DateCreated ??
+                product.PreviousTransition?.DateTransition ??
+                null
+            )
+          })}
+        </span>
+        {@html m.tasks_forNames({
+          allowedNames: product.ActiveTransition?.AllowedUserNames || m.appName(),
+          activityName: product.ActiveTransition?.InitialState ?? ''
+          // activityName appears to show up blank primarily at the very startup of a new product?
+        })}
+        {#if product.UserTasks.find((ut) => ut.UserId === page.data.session?.user.userId)}
+          <a class="link mx-2" href={localizeHref(`/tasks/${product.Id}`)}>
+            {m.common_continue()}
+          </a>
+        {/if}
+      {/if}
+    </div>
+  {/if}
+  <ProductDetails {product} transitions={product.Transitions} />
+</div>
